First versions of scripts to parse data from CHI and UIST, and add citations counts from Google.
The citation counts come from a KimonoLabs crawl of Google Search,
using the DOIs for the papers we're looking for.  The script scans the
results looking for citation counts and associated DOIs.  We also get
the publication title so that we can do some validation later with the
original dataset.

The other script just changes the format of the dataset, from a file
of values separated by tabs (tsv) to a JSON format.  The script also
builds the "referenced by" relationships, using the "cited by"
relationship and searching for the specified publications in the
dataset.