Changes script to extract citation counts to be more accurate.
The script now uses the URL associated with each result to determine
the DOI to which the result refers.  Before we were scanning the text
of the result for a DOI, but this often gave us a false match because
the ACM page includes DOIs for many different papers (citations and
references).