First implementation of PQ lookup function and namespace.
Adds a "paper" class that manages the lifecycle of a publication, it's
static and dynamic properties, and other helper functions to deal with
publications in general in the visualization.

Creates a PQ lookup function that should be used to interface with the
dataset.  This function manages a cache of paper objects and allows
convenient access to publications by their DOI.  It also exposes an
API for working with common sets of publications in the vis, like the
core papers, the fringe, etc.